\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{
  Suma dos números
}
\description{
  Suma dos números
}
\usage{
  suma.dos.numeros( a,b )
}

\arguments{
  \item{a}{ Número }
  \item{b}{ Número }
}
\details{
  Suma dos números}
\value{
A number.
}
\author{
Diego Andrés Ramírez Aragón
}
\examples{
# This command reads a few lines sampled from the EPA for the 1Q 2011
suma.dos.numeros(2,5)
}

\keyword{ manip }